4.21

+++ part4/blogList/controllers/blogs.js
@@ -70,6 +70,21 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+	const blog = await Blog.findById(request.params.id)
+	if (!blog) {
+		return response.status(404).end()
+	}
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
+
+	if (!user || blog.user.toString() !== user.id.toString()) {
+		return response.status(401).end()
+	}
+
 	await Blog.findByIdAndDelete(request.params.id)
 	response.status(204).end()
 })
